package/meson: make pkg-config --static aware
Our pkg-config wrapper is not used since commit
4e0bc29993376613d200e892d491e31ea5a49622, this raise static build
failures with libglib2 because --static is not passed anymore to
pkg-config so add a patch to get back the old behaviour.

The patch was proposed upstream
https://github.com/mesonbuild/meson/pull/6629, and the feedback was
not very positive. However, we need a solution that works for
Buildroot now, and this proposal is simple enough.

Fixes:

  http://autobuild.buildroot.org/results/0d36952def63cb69628697fa6408aeb6ce10cb5b

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit cf75d7da98596580eee5a9b5e1a3e156ad832099)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>